+++ lib/src/location_picker.dart
@@ -9,7 +9,8 @@ import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart' as intl;
 import 'package:latlong2/latlong.dart';
 
-import 'Widgets/wide_button.dart';
+import 'widgets/copyright_osm_widget.dart';
+import 'widgets/wide_button.dart';
 import 'classes.dart';
 
 /// Principal widget to show Flutter map using osm api with pick up location marker and search bar.
@@ -41,6 +42,10 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final String selectLocationButtonText;
 
+  /// [selectLocationButtonLeadingIcon] : (Widget) set the leading icon of the select location button
+  ///
+  final Widget? selectLocationButtonLeadingIcon;
+
   /// [initZoom] : (double) set initialized zoom in specific location  (default = 17)
   ///
   final double initZoom;
@@ -57,6 +62,10 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final double maxZoomLevel;
 
+  /// [maxBounds] : (LatLngBounds?) set default max bounds of the map (default = null)
+  ///
+  final LatLngBounds? maxBounds;
+
   /// [loadingWidget] : (Widget) show custom  widget until the map finish initialization
   ///
   final Widget? loadingWidget;
@@ -71,22 +80,18 @@ class FlutterLocationPicker extends StatefulWidget {
 
   /// [showZoomController] : (bool) enable/disable zoom in and zoom out buttons (default = true)
   ///
-
   final bool showZoomController;
 
   /// [showLocationController] : (bool) enable/disable locate me button (default = true)
   ///
-
   final bool showLocationController;
 
   /// [showSelectLocationButton] : (bool) enable/disable select location button (default = true)
   ///
-
   final bool showSelectLocationButton;
 
   /// [mapAnimationDuration] : (Duration) time duration of the move from point to point animation (default = Duration(milliseconds: 2000))
   ///
-
   final Duration mapAnimationDuration;
 
   /// [mapLoadingBackgroundColor] : (Color) change the background color of the loading screen before the map initialized
@@ -97,55 +102,83 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final ButtonStyle? selectLocationButtonStyle;
 
-  /// [selectLocationTextColor] : (Color) change the color of the select Location text
+  /// [selectLocationButtonWidth] : (double) change the width of the select Location button
   ///
-  final Color? selectLocationTextColor;
+  final double? selectLocationButtonWidth;
+
+  /// [selectLocationButtonHeight] : (double) change the height of the select Location button
+  ///
+  final double? selectLocationButtonHeight;
+
+  /// [selectedLocationButtonTextstyle] : set the style of the button text (default = TextStyle(fontSize: 20))
+  ///
+  final TextStyle selectedLocationButtonTextstyle;
+
+  /// [selectLocationButtonPositionTop] : (double) change the top position of the select Location button (default = null)
+  ///
+  final double? selectLocationButtonPositionTop;
+
+  /// [selectLocationButtonPositionRight] : (double) change the right position of the select Location button (default = 0)
+  ///
+  final double? selectLocationButtonPositionRight;
+
+  /// [selectLocationButtonPositionLeft] : (double) change the left position of the select Location button (default = 0)
+  ///
+  final double? selectLocationButtonPositionLeft;
+
+  /// [selectLocationButtonPositionBottom] : (double) change the bottom position of the select Location button (default = 3)
+  ///
+  final double? selectLocationButtonPositionBottom;
 
   /// [showSearchBar] : (bool) enable/disable search bar (default = true)
   ///
-
   final bool showSearchBar;
 
   /// [searchBarBackgroundColor] : (Color) change the background color of the search bar
   ///
-
   final Color? searchBarBackgroundColor;
 
   /// [searchBarTextColor] : (Color) change the color of the search bar text
   ///
-
   final Color? searchBarTextColor;
 
   /// [searchBarHintText] : (String) change the hint text of the search bar
   ///
-
   final String searchBarHintText;
 
   /// [searchBarHintColor] : (Color) change the color of the search bar hint text
   ///
-
   final Color? searchBarHintColor;
+
+  /// [searchbarInputBorder] : (OutlineInputBorder) change the border of the search bar
+  ///
+  final OutlineInputBorder? searchbarInputBorder;
+
+  /// [searchbarInputFocusBorder] : (OutlineInputBorder) change the border of the search bar when focused
+  ///
+  final OutlineInputBorder? searchbarInputFocusBorderp;
+
+  /// [searchbarBorderRadius] : (BorderRadiusGeometry) change the border radius of the search bar
+  ///
+  final BorderRadiusGeometry? searchbarBorderRadius;
 
   /// [zoomButtonsColor] : (Color) change the color of the zoom buttons icons
   ///
-
   final Color? zoomButtonsColor;
 
   /// [zoomButtonsBackgroundColor] : (Color) change the background color of the zoom buttons
   ///
-
   final Color? zoomButtonsBackgroundColor;
 
   /// [locationButtonsColor] : (Color) change the color of the location button icon
   ///
-
   final Color? locationButtonsColor;
 
   /// [locationButtonBackgroundColor] : (Color) change the background color of the location button
   ///
   final Color? locationButtonBackgroundColor;
 
-  /// [markerIcon] : (IconData) change the marker icon of the map (default = Icon(icons.location_on, color: Colors.red, size: 50))
+  /// [markerIcon] : (IconData) change the marker icon of the map (default = Icon(icons.location_on, color: Colors.blue, size: 50))
   ///
   final Widget? markerIcon;
 
@@ -153,15 +186,49 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final double markerIconOffset;
 
+  /// [showContributorBadgeForOSM] : (bool) for copyright of osm, we need to add badge in bottom of the map (default false)
+  ///
+  final bool showContributorBadgeForOSM;
+
+  /// [contributorBadgeForOSMColor] : (Color) change the color of the badge (default Colors.grey[300])
+  ///
+  final Color? contributorBadgeForOSMColor;
+
+  /// [contributorBadgeForOSMTextColor] : (Color) change the color of the badge text (default Colors.blue)
+  ///
+  final Color contributorBadgeForOSMTextColor;
+
+  /// [contributorBadgeForOSMText] : (String) change the text of the badge (default 'OpenStreetMap contributors')
+  ///
+  final String contributorBadgeForOSMText;
+
+  // [contributorBadgeForOSMPositionTop] : (double) change the position of the badge from top (default 0)
+  ///
+  final double? contributorBadgeForOSMPositionTop;
+
+  /// [contributorBadgeForOSMPositionLeft] : (double) change the position of the badge from left (default null)
+  ///
+  final double? contributorBadgeForOSMPositionLeft;
+
+  /// [contributorBadgeForOSMPositionRight] : (double) change the position of the badge from right (default 0)
+  ///
+  final double? contributorBadgeForOSMPositionRight;
+
+  /// [contributorBadgeForOSMPositionBottom] : (double) change the position of the badge from bottom (default -6)
+  ///
+  final double? contributorBadgeForOSMPositionBottom;
+
   const FlutterLocationPicker({
     Key? key,
     required this.onPicked,
+    this.selectedLocationButtonTextstyle = const TextStyle(fontSize: 20),
     this.onError,
     this.initPosition,
     this.stepZoom = 1,
     this.initZoom = 17,
     this.minZoomLevel = 2,
     this.maxZoomLevel = 18.4,
+    this.maxBounds,
     this.urlTemplate = 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
     this.mapLanguage = 'en',
     this.selectLocationButtonText = 'Set Current Location',
@@ -172,12 +239,20 @@ class FlutterLocationPicker extends StatefulWidget {
     this.showSelectLocationButton = true,
     this.showCurrentLocationPointer = true,
     this.selectLocationButtonStyle,
-    this.selectLocationTextColor,
+    this.selectLocationButtonWidth,
+    this.selectLocationButtonHeight,
+    this.selectLocationButtonPositionTop,
+    this.selectLocationButtonPositionRight = 0,
+    this.selectLocationButtonPositionLeft = 0,
+    this.selectLocationButtonPositionBottom = 3,
     this.showSearchBar = true,
     this.searchBarBackgroundColor,
     this.searchBarTextColor,
     this.searchBarHintText = 'Search location',
     this.searchBarHintColor,
+    this.searchbarInputBorder,
+    this.searchbarInputFocusBorderp,
+    this.searchbarBorderRadius,
     this.mapLoadingBackgroundColor,
     this.locationButtonBackgroundColor,
     this.zoomButtonsBackgroundColor,
@@ -185,7 +260,16 @@ class FlutterLocationPicker extends StatefulWidget {
     this.locationButtonsColor,
     this.markerIcon,
     this.markerIconOffset = 50.0,
+    this.showContributorBadgeForOSM = false,
+    this.contributorBadgeForOSMColor,
+    this.contributorBadgeForOSMTextColor = Colors.blue,
+    this.contributorBadgeForOSMText = 'OpenStreetMap contributors',
+    this.contributorBadgeForOSMPositionTop,
+    this.contributorBadgeForOSMPositionLeft,
+    this.contributorBadgeForOSMPositionRight = 0,
+    this.contributorBadgeForOSMPositionBottom = -6,
     Widget? loadingWidget,
+    this.selectLocationButtonLeadingIcon,
   })  : loadingWidget = loadingWidget ?? const CircularProgressIndicator(),
         super(key: key);
 
@@ -309,27 +393,18 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
   ///   longitude (double): The longitude of the location.
   void setNameCurrentPos(double latitude, double longitude) async {
     var client = http.Client();
-    setState(() {
-      isLoading = true;
-    });
     String url =
         'https://nominatim.openstreetmap.org/reverse?format=json&lat=$latitude&lon=$longitude&zoom=18&addressdetails=1&accept-language=${widget.mapLanguage}';
 
     try {
-      var response = await client.post(Uri.parse(url));
+      var response = await client.get(Uri.parse(url));
       var decodedResponse =
           jsonDecode(utf8.decode(response.bodyBytes)) as Map<dynamic, dynamic>;
       _searchController.text =
-          decodedResponse['display_name'] ?? "Search Location";
-      setState(() {
-        isLoading = false;
-      });
+          decodedResponse['display_name'] ?? "This Location is not accessible";
+      setState(() {});
     } on Exception catch (e) {
       onError(e);
-    } finally {
-      setState(() {
-        isLoading = false;
-      });
     }
   }
 
@@ -344,12 +419,20 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
     var client = http.Client();
     String url =
         'https://nominatim.openstreetmap.org/reverse?format=json&lat=${_mapController.center.latitude}&lon=${_mapController.center.longitude}&zoom=18&addressdetails=1&accept-language=${widget.mapLanguage}';
-
-    var response = await client.post(Uri.parse(url));
+    var response = await client.get(Uri.parse(url));
     var decodedResponse =
         jsonDecode(utf8.decode(response.bodyBytes)) as Map<dynamic, dynamic>;
-    String displayName = decodedResponse['display_name'];
-    return PickedData(center, displayName, decodedResponse['address']);
+    String displayName = "This Location is not accessible";
+    Map<String, dynamic> address;
+
+    if (decodedResponse['display_name'] != null) {
+      displayName = decodedResponse['display_name'];
+      address = decodedResponse['address'];
+    } else {
+      center = const LatLong(0, 0);
+      address = decodedResponse as Map<String, dynamic>;
+    }
+    return PickedData(center, displayName, address);
   }
 
   @override
@@ -459,7 +542,8 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
         decoration: BoxDecoration(
           color: widget.searchBarBackgroundColor ??
               Theme.of(context).colorScheme.background,
-          borderRadius: BorderRadius.circular(5),
+          borderRadius:
+              widget.searchbarBorderRadius ?? BorderRadius.circular(5),
         ),
         child: Column(
           children: [
@@ -475,11 +559,16 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
                   hintTextDirection: isRTL(widget.searchBarHintText)
                       ? TextDirection.rtl
                       : TextDirection.ltr,
-                  border: inputBorder,
-                  focusedBorder: inputFocusBorder,
+                  border: widget.searchbarInputBorder ?? inputBorder,
+                  focusedBorder:
+                      widget.searchbarInputFocusBorderp ?? inputFocusBorder,
                   hintStyle: TextStyle(color: widget.searchBarHintColor),
                   suffixIcon: IconButton(
-                    onPressed: () => _searchController.clear(),
+                    onPressed: () {
+                      _searchController.clear();
+                      _options.clear();
+                      setState(() {});
+                    },
                     icon: Icon(
                       Icons.clear,
                       color: widget.searchBarTextColor,
@@ -496,7 +585,7 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
                     try {
                       String url =
                           'https://nominatim.openstreetmap.org/search?q=$value&format=json&polygon_geojson=1&addressdetails=1&accept-language=${widget.mapLanguage}';
-                      var response = await client.post(Uri.parse(url));
+                      var response = await client.get(Uri.parse(url));
                       var decodedResponse =
                           jsonDecode(utf8.decode(response.bodyBytes))
                               as List<dynamic>;
@@ -564,9 +653,9 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
               heroTag: "btn3",
               backgroundColor: widget.locationButtonBackgroundColor,
               onPressed: () async {
-                setState(() {
-                  isLoading = true;
-                });
+                // setState(() {
+                //   isLoading = true;
+                // });
                 _determinePosition().then((currentPosition) {
                   _animatedMapMove(
                       LatLng(
@@ -588,10 +677,12 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
     return Positioned.fill(
         child: FlutterMap(
       options: MapOptions(
-          center: initPosition,
-          zoom: initPosition != null ? widget.initZoom : widget.minZoomLevel,
-          maxZoom: widget.maxZoomLevel,
-          minZoom: widget.minZoomLevel),
+        center: initPosition,
+        zoom: initPosition != null ? widget.initZoom : widget.minZoomLevel,
+        maxZoom: widget.maxZoomLevel,
+        minZoom: widget.minZoomLevel,
+        maxBounds: widget.maxBounds,
+      ),
       mapController: _mapController,
       children: [
         TileLayer(
@@ -624,7 +715,7 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
           child: widget.markerIcon ??
               const Icon(
                 Icons.location_pin,
-                color: Colors.red,
+                color: Colors.blue,
                 size: 50,
               ),
         ),
@@ -634,25 +725,31 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
 
   Widget _buildSelectButton() {
     return Positioned(
-      bottom: 0,
-      left: 0,
-      right: 0,
+      top: widget.selectLocationButtonPositionTop,
+      bottom: widget.selectLocationButtonPositionBottom,
+      left: widget.selectLocationButtonPositionLeft,
+      right: widget.selectLocationButtonPositionRight,
       child: Center(
         child: Padding(
           padding: const EdgeInsets.all(8.0),
-          child: WideButton(widget.selectLocationButtonText,
-              onPressed: () async {
-            setState(() {
-              isLoading = true;
-            });
-            pickData().then((value) {
-              widget.onPicked(value);
-            }, onError: (e) => onError(e)).whenComplete(() => setState(() {
-                  isLoading = false;
-                }));
-          },
-              style: widget.selectLocationButtonStyle,
-              textColor: widget.selectLocationTextColor),
+          child: WideButton(
+            widget.selectLocationButtonText,
+            leadingIcon: widget.selectLocationButtonLeadingIcon,
+            onPressed: () async {
+              setState(() {
+                isLoading = true;
+              });
+              pickData().then((value) {
+                widget.onPicked(value);
+              }, onError: (e) => onError(e)).whenComplete(() => setState(() {
+                    isLoading = false;
+                  }));
+            },
+            style: widget.selectLocationButtonStyle,
+            textStyle: widget.selectedLocationButtonTextstyle,
+            width: widget.selectLocationButtonWidth,
+            height: widget.selectLocationButtonHeight,
+          ),
         ),
       ),
     );
@@ -660,17 +757,32 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
 
   @override
   Widget build(BuildContext context) {
-    return SafeArea(
-      child: Stack(
-        children: [
-          _buildMap(),
-          if (!isLoading) _buildMarker(),
-          if (isLoading) Center(child: widget.loadingWidget!),
-          _buildControllerButtons(),
-          if (widget.showSearchBar) _buildSearchBar(),
-          if (widget.showSelectLocationButton) _buildSelectButton()
-        ],
-      ),
+    return Stack(
+      children: [
+        _buildMap(),
+        if (!isLoading) _buildMarker(),
+        if (isLoading) Center(child: widget.loadingWidget!),
+        SafeArea(
+          child: Stack(children: [
+            _buildControllerButtons(),
+            if (widget.showSearchBar) _buildSearchBar(),
+            if (widget.showContributorBadgeForOSM) ...[
+              Positioned(
+                top: widget.contributorBadgeForOSMPositionTop,
+                bottom: widget.contributorBadgeForOSMPositionBottom,
+                left: widget.contributorBadgeForOSMPositionLeft,
+                right: widget.contributorBadgeForOSMPositionRight,
+                child: CopyrightOSMWidget(
+                  badgeText: widget.contributorBadgeForOSMText,
+                  badgeTextColor: widget.contributorBadgeForOSMTextColor,
+                  badgeColor: widget.contributorBadgeForOSMColor,
+                ),
+              ),
+            ],
+            if (widget.showSelectLocationButton) _buildSelectButton(),
+          ]),
+        )
+      ],
     );
   }
 }